mwfertiliserpilot: Fertilizer management pilot survey for maize cultivation in Ntonda EPA, Malawi

In July 2024, 129 farmers were surveyed on fertilizer management for maize cultivation in the Ntonda Extension Planning Area (EPA) of Blantyre District, Southern Region of Malawi.

Usage

mwfertiliserpilot

Format

A tibble with 129 rows and 123 variables

start

Survey start

end

Survey end

today

Survey date

consent

Do you give consent to proceed with this survey?

village_name

Please enter the name of the village.

gvh_name

Please enter the name of the group village head (GVH).

ta_name

Please enter the name of the traditional authority (TA).

farmer_gender

Please enter the farmer's gender.

farmer_age

Please enter the farmer's age.

farmer_hoh

Are you the head of the household?

hoh_education

What is your highest level of education?

n_years_farming

How many years have you been farming?

own_livestock

Do you own any livestock (cows, chickens, pigs, etc.)?

own_livestock_animals

Do you own...?

own_livestock_animals.cows

Cows

own_livestock_animals.chickens

Chickens

own_livestock_animals.pigs

Pigs

own_livestock_animals.goats

Goats

own_livestock_animals.sheep

Sheep

own_livestock_animals.other

Other livestock

farming_main_income

Is farming your household's main source of income?

household_main_income

What is your household's main source of income?

n_acres_farm

How large is your farm (in acres)?

n_acres_maize

How much area is dedicated to maize farming (in acres)?

source_water_farm

Is this area rain-fed or part of an irrigation scheme?

chem_fertilizer

Did you apply chemical fertilizer this year?

chem_fertilizer_type

What chemical fertilizer do you use?

n_bags_npk

How many bags of NPK did you apply THIS YEAR?

n_bags_npk_subsidized

How many of those were subsidised?

price_bags_npk_nonsubsidized

How much did you pay for a full-priced bag of NPK?

price_bags_npk_subsidized

How much did you pay for a subsidised bag of NPK?

enough_bags_npk

Was this enough NPK to meet your needs?

n_bags_npk_desired

How many bags of NPK would you have liked?

n_bags_npk_buying_next_year

How many bags of NPK will you try to buy for the coming year?

n_bags_urea

How many bags of urea did you apply THIS YEAR?

n_bags_urea_subsidized

How many of those were subsidised?

price_bags_urea_nonsubsidized

How much did you pay for a full-priced bag of urea?

price_bags_urea_subsidized

How much did you pay for a subsidised bag of urea?

enough_bags_urea

Was this enough urea to meet your needs?

n_bags_urea_desired

How many bags of urea would you have liked?

n_bags_urea_buying_next_year

How many bags of urea will you try to buy for the coming year?

able_buy_bags_past

Were you able to buy all the fertilizer you require 5 years ago?

n_bags_maize_t1_season

How many bags of maize did you harvest in the last growing season?

n_bags_maize_t2_season

How many bags of maize did you harvest two growing seasons ago?

n_bags_maize_t3_season

How many bags of maize did you harvest three growing seasons ago?

fertilizer_comparison_5years

Do you need more, less or the same amount of chemical fertilizer to grow the same amount of maize now as you did 5 years ago?

same_tool_for_fertilizer

Do you use the same tool to apply NPK (basal) and urea (topdressing)?

single_tool_fertilizer

What tool do you use to apply them?

single_tool_fertilizer_other

Please specify.

tool_basal_fertilizer

What tool do you use to apply NPK (basal fertilizer) in your field?

tool_basal_fertilizer_other

Please specify.

tool_topdressing_fertilizer

What tool do you use to apply urea (topdressing fertilizer) in your field?

tool_topdressing_fertilizer_other

Please specify.

timing_fertilizer

Do you apply NPK and urea at different times or the same time?

timing_fertilizer_mix_ntimes

Do you apply the mix once or twice?

timing_fertilizer_mix

When do you apply them?

timing_fertilizer_mix_other

Please specify.

place_fertilizer_mix

Where do you apply them?

place_fertilizer_mix_other

Please specify.

ratio_fertilizer_mix_npk

What's the NPK to urea ratio? Enter the number for NPK:

ratio_fertilizer_mix_urea

What's the NPK to urea ratio? Enter the number for urea:

timing_fertilizer_mix_t1

When do you apply them for the first application?

timing_fertilizer_mix_t1_other

Please specify.

place_fertilizer_mix_t1

Where do you apply them for the first application?

place_fertilizer_mix_t1_other

Please specify.

ratio_fertilizer_mix_t1_npk

What's the NPK to urea ratio for the first application? Enter the number for NPK:

ratio_fertilizer_mix_t1_urea

What's the NPK to urea ratio for the first application? Enter the number for urea:

timing_fertilizer_mix_t2

When do you apply them for the second application?

timing_fertilizer_mix_t2_other

Please specify.

place_fertilizer_mix_t2

Where do you apply them for the second application?

place_fertilizer_mix_t2_other

Please specify.

ratio_fertilizer_mix_t2_npk

What's the NPK to urea ratio for the second application? Enter the number for NPK:

ratio_fertilizer_mix_t2_urea

What's the NPK to urea ratio for the second application? Enter the number for urea:

timing_basal_fertilizer

When do you apply NPK (basal fertilizer)?

timing_basal_fertilizer_other

Please specify.

place_basal_fertilizer

Where do you apply NPK (basal fertilizer)?

place_basal_fertilizer_other

Please specify.

timing_topdressing

When do you apply urea (topdressing fertilizers)?

timing_topdressing_other

Please specify.

place_topdressing

Where do you apply urea (topdressing fertilizers)?

place_topdressing_other

Please specify.

basal_fertilizer_grams

Grams of basal fertilizer applied

topdressing_fertilizer_grams

Grams of topdressing fertilizer applied

know_recommended_tools

Do you know the recommended cups (cups #5 or #8) for maize?

reason_no_use_scoop

Why are you not using them?

reason_no_use_scoop_other

Please elaborate.

would_use_scoop_if_given

Would you use the fertilizer scoops if they were given to you?

would_use_scoop_if_available

Would you buy the fertilizer scoops if available?

wtp_scoop

How much would you be willing to pay for the fertilizer scoops?

selling_points_scoop

Do you know where the scoops are sold?

selling_points_scoop_location

Can you state where?

know_price_scoop

Do you know how much it costs in the shop?

price_scoop

How much?

type_seed

What type of seed do you use?

n_seeds

How many seeds do you plant per planting station?

planting_distance

Measure how far the planting stations are away from each other.

ridges_distance

Measure how far the ridges are away from each other.

several_plants_emerging

If multiple plants emerge from the same planting station, select "Yes".

intercropping

Do you practice intercropping?

intercrop_type

What do you intercrop with?

intercrop_type_other

Please specify.

pesticide_use

Do you use pesticides?

fertilizer_subsidy_received

Were you a beneficiary of fertilizer subsidies this year?

concern_fertilizer_removed

How concerned are you about fertilizer subsidies being removed?

concern_fertilizer_removed_explain

Please explain.

fertilizer_programme_effective

Do you think the fertilizer subsidy programme reaches the people it is meant to?

fertilizer_programme_improvements

Do you think the fertilizer subsidy programme could be improved?

fertilizer_programme_improvements_explain

Please explain.

apply_org_fertilizer

Do you apply organic fertilizers to your fields?

origin_org_fertilizer

Where are you getting the organic fertilizer from?

how_apply_org_fertilizer

How do you apply it?

challenges_org_fertilizer

Do you have any challenges with organic fertilizers?

type_challenges_org_fertilizer

What challenges do you have?

type_challenges_org_fertilizer_other

Please elaborate.

sufficient_org_fertilizer

Do you think organic fertilizers would be sufficient for your fields without any chemical fertilizer?

subsidies_org_fertilizer

Should the government subsidise organic fertilizers?

health_soil_now

How healthy do you think your soils are?

health_soil_t10

How healthy do you think your soils were 10 years ago?

reasons_changes

Why do you think so?

suggestions_soil_improvement

What do you think should be done to restore/maintain/improve the soil health in your field?

final_comments

Do you have anything else you would like to tell us?

end_message

Thank you for your time. The survey is now finished. Thank the person for their availability, time and answers. Ask if they have any further questions. This space allows you to report anything that may have been unusual during the survey.

uuid

Unique identifier